DRY logic to extract scripture
